+++ R/templates.R
@@ -1,19 +1,36 @@
 #' @include S4classes.R
 NULL
 
-#' Get template for reconstructing full text.
+#' Get template for formatting full text output.
 #' 
-#' Templates are used to format the markdown/html output of partitions.
+#' Templates are used to format the markdown/html of fulltext. `get_template()`
+#' loads and parses the content of the JSON template file, if it is found in the
+#' slot "template" of `.Object`.
 #' 
-#' @param .Object A \code{corpus}, \code{subcorpus} or \code{partition} object,
-#'   or a length-one \code{character} vector with a corpus ID.
-#' @param warn A \code{logical} value, whether to issue a warning if template 
-#'   is not available. Defaults to \code{FALSE}.
-#' @param ... Further arguments to be defined.
+#' To learn how to write templates, consult the sample files in the folder
+#' "templates" of the installed package - see examples.
+#' @details 
+#' @param .Object A `corpus`, `subcorpus` or `partition` object,
+#'   or a length-one `character` vector with a corpus ID.
+#' @param warn A `logical` value, whether to issue a warning if template 
+#'   is not available. Defaults to `FALSE`.
+#' @param ... Further arguments.
+#' @return If a template is available, a `list` with the parsed content of the
+#'   template file, otherwise `NULL`.
 #' @exportMethod get_template
 #' @rdname templates
-#' @aliases get_template,subcorpus-method
-setGeneric("get_template", function(.Object, ...) standardGeneric("get_template"))
+#' @examples
+#' use("polmineR")
+#' corpus("GERMAPARLMINI") %>%
+#'   get_template()
+#'
+#' # template files included in the package   
+#' template_dir <- system.file(package = "polmineR", "templates")
+#' list.files(template_dir)
+setGeneric(
+  "get_template",
+  function(.Object, ...) standardGeneric("get_template")
+)
 
 #' @rdname templates
 setMethod("get_template", "character", function(.Object, warn = FALSE){
@@ -23,8 +40,8 @@ setMethod("get_template", "character", function(.Object, warn = FALSE){
 #' @rdname templates
 setMethod("get_template", "corpus", function(.Object, warn = FALSE){
   if (is.na(.Object@template)){
-    if (warn) warning(
-      sprintf("No template available for corpus '%s'.", .Object@corpus)
+    if (warn) cli_alert_warning(
+      "No template available for corpus {.val {.Object@corpus}}."
     )
     return(NULL)
   } else {